backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
PyJWT==2.8.0
email-validator==2.1.0
bcrypt==4.1.1
passlib==1.7.4